# docs/API/openapi.yaml
openapi: 3.1.0
info:
  title: School Management System API
  version: 1.1.0
  description: API for multi-tenant School Management System
servers:
  - url: https://api.example.com/api/v1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    PageParam:
      in: query
      name: page
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    SortParam:
      in: query
      name: sort
      description: Sort expression (e.g., "name,-createdAt")
      schema:
        type: string
...
  schemas:
...
    Meta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        period: { type: string }
        dueDate: { type: string, format: date }
        amount: { type: integer, description: Amount in minor units }
        status: { type: string, enum: [draft, issued, paid, partial, overdue, void] }
...
paths:
...
  /fees/invoices:
    get:
      summary: List invoices
      tags: [Fees]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Invoice' }
                  meta: { $ref: '#/components/schemas/Meta' }
    post:
      summary: Create invoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentIds, period, dueDate, lines]
              properties:
                studentIds:
                  type: array
                  items: { type: string, format: uuid }
                period: { type: string }
                dueDate: { type: string, format: date }
                lines:
                  type: array
                  items:
                    type: object
                    required: [feeHeadId, amount]
                    properties:
                      feeHeadId: { type: string, format: uuid }
                      amount: { type: integer }
      responses:
        '200': { description: OK }
...
  /fees/invoices/{id}:
    get:
      summary: Get invoice by id
      operationId: getInvoice
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
...
  /fees/invoices/{id}/export:
    post:
      summary: Export invoice as PDF and store in S3
      operationId: exportInvoicePdf
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  key: { type: string }
                  url: { type: string }
...